requirements.txt
streamlit
pytesseract
Pillow
pandas
matplotlib
langchain
openai
python-dotenv
